% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{add_node.prepper}
\alias{add_node.prepper}
\title{Add a pre-processing node to a pipeline}
\usage{
\method{add_node}{prepper}(x, step, ...)
}
\arguments{
\item{x}{A \code{prepper} pipeline}

\item{step}{The pre-processing step to add}

\item{...}{Additional arguments}
}
\description{
Add a pre-processing node to a pipeline
}
