% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{add_node}
\alias{add_node}
\title{add a pre-processing stage}
\usage{
add_node(x, step, ...)
}
\arguments{
\item{x}{the processing pipeline}

\item{step}{the pre-processing step to add}

\item{...}{extra args}
}
\value{
a new pre-processing pipeline with the added step
}
\description{
add a pre-processing stage
}
