% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{apply_transform}
\alias{apply_transform}
\title{apply a pre-processing transform}
\usage{
apply_transform(x, X, colind, ...)
}
\arguments{
\item{x}{the pre_processor}

\item{X}{the data matrix}

\item{colind}{column indices}

\item{...}{extra args}
}
\value{
the transformed data
}
\description{
apply a pre-processing transform
}
