% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_projector_union.R
\name{bi_projector_union}
\alias{bi_projector_union}
\title{a union of concatenated \code{bi_projector} fits}
\usage{
bi_projector_union(fits, outer_block_indices = NULL)
}
\arguments{
\item{fits}{a list of \code{bi_projector} instances with the same row space.}

\item{outer_block_indices}{list of indices for the outer blocks}
}
\description{
given a set of \code{bi_projector} fits, join the together to create a new \code{bi_projector} instance.
The new weights and associated scores will simply be concatenated.
}
\examples{

X1 <- matrix(rnorm(5*5), 5,5)
X2 <- matrix(rnorm(5*5), 5,5)

bpu <- bi_projector_union(list(pca(X1), pca(X2)))

}
