% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{block_indices.multiblock_projector}
\alias{block_indices.multiblock_projector}
\title{Extract the Block Indices from a Multiblock Projector}
\usage{
\method{block_indices}{multiblock_projector}(x, i, ...)
}
\arguments{
\item{x}{A \code{multiblock_projector} object.}

\item{i}{Ignored.}

\item{...}{Ignored.}
}
\value{
The list of block indices.
}
\description{
Extract the Block Indices from a Multiblock Projector
}
