% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_indices}
\alias{block_indices}
\title{get block_indices}
\usage{
block_indices(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{extra args}
}
\value{
a list of block indices
}
\description{
extract the list of indices associated with each block in a \code{multiblock} object
}
