% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_lengths}
\alias{block_lengths}
\title{get block_lengths}
\usage{
block_lengths(x)
}
\arguments{
\item{x}{the object}
}
\value{
the block lengths
}
\description{
extract the lengths of each block in a multiblock object
}
