% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-utils.R
\name{check_fitted}
\alias{check_fitted}
\title{Check if preprocessor is fitted and error if not}
\usage{
check_fitted(object, action = "transform")
}
\arguments{
\item{object}{A preprocessing object}

\item{action}{Character string describing the attempted action}
}
\description{
Internal helper to provide consistent error messages when
attempting to transform with unfitted preprocessors.
}
\keyword{internal}
