% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.discriminant_projector}
\alias{classifier.discriminant_projector}
\title{Create a k-NN classifier for a discriminant projector}
\usage{
\method{classifier}{discriminant_projector}(x, colind = NULL, knn = 1, ...)
}
\arguments{
\item{x}{the discriminant projector object}

\item{colind}{an optional vector specifying the column indices of the components}

\item{knn}{the number of nearest neighbors (default=1)}

\item{...}{extra arguments}
}
\value{
a classifier object
}
\description{
Create a k-NN classifier for a discriminant projector
}
\examples{
# Assume dp is a fitted discriminant_projector object
# classifier(dp, knn = 5) # Basic example
}
