% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.multiblock_biprojector}
\alias{classifier.multiblock_biprojector}
\title{Multiblock Bi-Projector Classifier}
\usage{
\method{classifier}{multiblock_biprojector}(
  x,
  colind = NULL,
  labels,
  new_data = NULL,
  block = NULL,
  global_scores = TRUE,
  knn = 1,
  ...
)
}
\arguments{
\item{x}{A fitted multiblock bi-projector model object.}

\item{colind}{An optional vector of column indices used for prediction (default: NULL).}

\item{labels}{A factor or vector of class labels for the training data.}

\item{new_data}{An optional data matrix for which to generate predictions (default: NULL).}

\item{block}{An optional block index for prediction (default: NULL).}

\item{global_scores}{Whether to use the global scores or the partial scores for reference space (default: TRUE).}

\item{knn}{The number of nearest neighbors to consider in the classifier (default: 1).}

\item{...}{Additional arguments.}
}
\value{
A multiblock classifier object.
}
\description{
Constructs a classifier for a multiblock bi-projector model object.
Either global or partial scores can be used. If \code{colind} or \code{block} are provided
and \code{global_scores=FALSE}, partial projection is performed. Otherwise, global projection is used.
}
\concept{classifier}
