% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/classifier.R
\name{classifier}
\alias{classifier}
\alias{classifier.projector}
\title{construct a classifier}
\usage{
classifier(x, colind, ...)

\method{classifier}{projector}(x, labels, new_data, colind = NULL, knn = 1)
}
\arguments{
\item{x}{the model object}

\item{colind}{the (optional) column indices used for prediction}

\item{...}{extra arguments}

\item{labels}{the labels associated with the rows of the projected data (see \code{new_data})}

\item{new_data}{reference data associated with \code{labels} and to be projected into subspace (required).}

\item{knn}{the number of nearest neighbors to use when classifying a new point.}
}
\description{
Given a model object (e.g. \code{projector} construct a classifier that can generate predictions for new data points.
}
\examples{
data(iris)
X <- iris[,1:4]
pcres <- pca(as.matrix(X),2)
cfier <- classifier(pcres, iris[,5], new_data=iris[,1:4])
}
