% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed_projector.R
\name{coef.composed_projector}
\alias{coef.composed_projector}
\title{Get Coefficients of a Composed Projector}
\usage{
\method{coef}{composed_projector}(object, ...)
}
\arguments{
\item{object}{A \code{composed_projector} object.}

\item{...}{Currently unused.}
}
\value{
A matrix representing the combined coefficients.
}
\description{
Calculates the effective coefficient matrix that maps from the original
input space (of the first projector) to the final output space (of the
last projector). This is done by multiplying the coefficient matrices
of all projectors in the sequence.
}
