% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{coef.multiblock_projector}
\alias{coef.multiblock_projector}
\title{Coefficients for a Multiblock Projector}
\usage{
\method{coef}{multiblock_projector}(object, block, ...)
}
\arguments{
\item{object}{A \code{multiblock_projector} object.}

\item{block}{Optional block index. If missing, returns loadings for all variables.}

\item{...}{Additional arguments.}
}
\value{
A matrix of loadings.
}
\description{
Extracts the components (loadings) for a given block or the entire projector.
}
