% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{coef.twoway_projector}
\alias{coef.twoway_projector}
\title{get coefficients}
\usage{
\method{coef}{twoway_projector}(x, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{the model fit}

\item{source}{the source of the data (X or Y block)}
}
\description{
extract X or Y coefficents for two-way projector
}
