% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{colscale}
\alias{colscale}
\title{scale a data matrix}
\usage{
colscale(preproc = prepper(), type = c("unit", "z", "weights"), weights = NULL)
}
\description{
normalize each column by a scale factor.
}
