% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed_projector.R
\name{compose_projectors}
\alias{compose_projectors}
\title{projector composition}
\usage{
compose_projectors(...)
}
\arguments{
\item{...}{the sequence of projectors}
}
\value{
a class of type \code{composed_projector} that extends \code{function}
}
\description{
compose a sequence of \code{projector}s in forward order
}
\examples{
 
 X <- matrix(rnorm(20*20), 20, 20)
 pca1 <- pca(X, ncomp=10)
 X2 <- scores(pca1)
 pca2 <- pca(X2, ncomp=4)
 
 cproj <- compose_projectors(pca1,pca2)

 stopifnot(ncol(cproj(X)) == 4)
 all.equal(project(cproj, X),cproj(X))
 
}
