% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_projector.R
\name{discriminant_projector}
\alias{discriminant_projector}
\title{construct a discriminant projector}
\usage{
discriminant_projector(
  v,
  s,
  sdev,
  preproc = prep(pass()),
  labels,
  classes = NULL,
  ...
)
}
\arguments{
\item{v}{a matrix of coefficients with dimension \code{nrow(v)} by \code{ncol(v)} (number of columns = number of components)}

\item{s}{The score matrix}

\item{sdev}{The standard deviations of the score matrix}

\item{preproc}{(optional) A pre-processing pipeline, default is prep(pass())}

\item{labels}{the training labels}

\item{classes}{(optional) A character vector specifying the class attributes of the object, default is NULL}

\item{...}{extra args}
}
\description{
a \code{discriminant_projector} instance extending \code{bi_projeector} whose projection maximizes class separation.
}
