% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_operator.R
\name{effect_operator}
\alias{effect_operator}
\title{Construct an effect operator}
\usage{
effect_operator(v, s, sdev, preproc = fit(pass(), matrix(0, 1, nrow(v))), ...)
}
\arguments{
\item{v}{Feature loadings in original variable space.}

\item{s}{Observation-side effect scores.}

\item{sdev}{Singular values.}

\item{preproc}{Fitted response preprocessor.}

\item{...}{Additional metadata stored on the object.}
}
\value{
An \code{effect_operator}.
}
\description{
Construct an effect operator
}
