% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{feature_importance.classifier}
\alias{feature_importance.classifier}
\title{Evaluate Feature Importance}
\usage{
\method{feature_importance}{classifier}(
  x,
  new_data,
  ncomp = NULL,
  blocks = NULL,
  metric = c("cosine", "euclidean", "ejaccard"),
  fun = rank_score,
  normalize_probs = FALSE,
  approach = c("marginal", "standalone"),
  ...
)
}
\arguments{
\item{x}{classifier}

\item{new_data}{new data}

\item{ncomp}{...}

\item{blocks}{a list of feature indices}

\item{metric}{...}

\item{fun}{a function to compute accuracy (default rank_score)}

\item{normalize_probs}{logical}

\item{approach}{"marginal" or "standalone"}

\item{...}{args to projection}
}
\value{
a data.frame with block and importance
}
\description{
Uses "marginal" or "standalone" approaches:
\itemize{
\item marginal: remove block and see change in accuracy
\item standalone: use only that block and measure accuracy
}
}
