% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{feature_importance}
\alias{feature_importance}
\title{Evaluate feature importance}
\usage{
feature_importance(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\value{
the feature importance scores
}
\description{
Calculate the importance of features in a model
}
