% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{fresh.prepper}
\alias{fresh.prepper}
\title{Create a fresh pipeline from an existing prepper}
\usage{
\method{fresh}{prepper}(x, ...)
}
\description{
Recreates the pipeline structure without any learned parameters.
}
