% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{fresh}
\alias{fresh}
\title{Get a fresh pre-processing node cleared of any cached data}
\usage{
fresh(x, ...)
}
\arguments{
\item{x}{the processing pipeline}

\item{...}{extra args}
}
\value{
a fresh pre-processing pipeline
}
\description{
Get a fresh pre-processing node cleared of any cached data
}
