% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneig.R
\name{geneig}
\alias{geneig}
\title{generalized eigenvalue decomposition}
\usage{
geneig(A, B, ncomp = 2, method = c("robust", "sdiag", "lapack"))
}
\arguments{
\item{A}{the left hand side matrix}

\item{B}{the right hand side matrix}

\item{ncomp}{number of components to return}

\item{method}{robust or lapack (using \code{geigen} package)}

\item{ordering}{the ordering of the eigenvalues/vectors (LM = largest magnitude first, SM = smallest magnitude first)}
}
\value{
\code{geneig} instance whcih is a subclass of \code{projector} with added slot for eigenvalues called \code{values}
}
\description{
compute generalized eigenvalues and eigenvectors using one of two methods
}
\examples{

A <- matrix(c(14, 10, 12,
              10, 12, 13,
              12, 13, 14), nrow=3, byrow=TRUE)
B <- matrix(c(48, 17, 26,
              17, 33, 32,
              26, 32, 34), nrow=3, byrow=TRUE)
              
}
