% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-utils.R
\name{get_fitted_state}
\alias{get_fitted_state}
\title{Get fitted state from attributes}
\usage{
get_fitted_state(object)
}
\arguments{
\item{object}{A preprocessing object}
}
\value{
Logical indicating fitted state, or NULL if not tracked
}
\description{
Get fitted state from attributes
}
\keyword{internal}
