% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_means}
\alias{group_means}
\title{compute column-wise mean in \code{X} for each factor level of \code{Y}}
\usage{
group_means(Y, X)
}
\arguments{
\item{Y}{a vector of labels to compute means over disjoint sets}

\item{X}{a data matrix from which to compute means}
}
\description{
compute group means
}
