% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_utils.R
\name{identity_basis}
\alias{identity_basis}
\title{Identity basis specification}
\usage{
identity_basis()
}
\value{
A basis specification for the unregularized feature space.
}
\description{
Identity basis specification
}
