% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{init_transform}
\alias{init_transform}
\title{initialize a transform}
\usage{
init_transform(x, X, ...)
}
\arguments{
\item{x}{the pre_processor}

\item{X}{the data matrix}
}
\value{
an initialized pre-processor
}
\description{
initialize a transform
}
\keyword{internal}
