% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{inverse_projection}
\alias{inverse_projection}
\title{inverse of the component matrix}
\usage{
inverse_projection(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
return the inverse projection matrix. Can be used to map back to data space.
If the component matrix is orthogonal, then the inverse projection is the transpose of the component matrix.
}
