% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-utils.R
\name{is_fitted}
\alias{is_fitted}
\title{Check if a preprocessing object is fitted}
\usage{
is_fitted(object)
}
\arguments{
\item{object}{A preprocessing object to check}
}
\value{
Logical: TRUE if fitted, FALSE otherwise
}
\description{
Determine whether a preprocessing object has been fitted to data.
This is used internally to provide helpful error messages when
users try to transform data with an unfitted preprocessor.
}
\keyword{internal}
