% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{is_orthogonal.projector}
\alias{is_orthogonal.projector}
\title{Stricter check for true orthogonality}
\usage{
\method{is_orthogonal}{projector}(x, tol = 1e-06)
}
\arguments{
\item{x}{the projector object}

\item{tol}{tolerance for checking orthogonality}
}
\description{
We test if v^T * v = I (when rows >= cols) or v * v^T = I (when cols > rows).
}
