% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{is it orthogonal}
\usage{
is_orthogonal(x)
}
\arguments{
\item{x}{the object}
}
\value{
a logical value indicating whether the transformation is orthogonal
}
\description{
test whether components are orthogonal
}
