% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{multiframe}
\alias{multiframe}
\title{create an multiframe object}
\usage{
multiframe(x, y, ...)
}
\arguments{
\item{x}{the multivariate data (a matrix, a list, or other data container)}

\item{y}{a design matrix with same number of rows/elements as x}

\item{...}{extra args}
}
\description{
construct a new multivariate design object linking vector-valued observation and arbitrary design variables
}
\examples{

X <- matrix(rnorm(20*100), 20, 100)
Y <- tibble(condition=rep(letters[1:5], 4))

mds <- multidesign(X,Y)
sdes <- split(mds, condition)
}
