% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{nblocks}
\alias{nblocks}
\title{get the number of blocks}
\usage{
nblocks(x)
}
\arguments{
\item{x}{the object}
}
\value{
the number of blocks
}
\description{
The number of data blocks in a multiblock element
}
