% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{new_classifier}
\alias{new_classifier}
\title{constrcut a new classifier}
\usage{
new_classifier(x, labels, scores, colind = NULL, knn = 1, classes = NULL, ...)
}
\arguments{
\item{x}{the model fit}

\item{labels}{the class labels}

\item{colind}{the column indices}

\item{knn}{the number of nearest neighbors for prediction}

\item{classes}{addiitonal s3 classes}

\item{...}{extra args}
}
\description{
constrcut a new classifier
}
\keyword{internal}
