% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{observation}
\alias{observation}
\title{create an observation object}
\usage{
observation(x, i)
}
\arguments{
\item{x}{the data source}

\item{i}{the index of the observation}
}
\description{
construct a new multivariate observation vector
}
