% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_project}
\alias{partial_project}
\title{partially project a new sample onto subspace}
\usage{
partial_project(x, new_data, colind)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of columns as the original data.}

\item{colind}{the column indices to select in the projection matrix}
}
\description{
project a selected subset of column indices onto the subspace
}
