% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{partial_projector.projector}
\alias{partial_projector.projector}
\title{construct a partial_projector from a \code{projector} instance}
\usage{
\method{partial_projector}{projector}(x, colind, ...)
}
\description{
construct a partial_projector from a \code{projector} instance
}
\examples{

X <- matrix(rnorm(10*10), 10, 10)
pfit <- pca(X, ncomp=9)
proj <- project(pfit, X)

pp <- partial_projector(pfit, 1:5)
}
