% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_projector}
\alias{partial_projector}
\title{construct a partial projector}
\usage{
partial_projector(x, colind, ...)
}
\arguments{
\item{x}{the original \code{projector} instance}

\item{colind}{the column indices to select in the projection matrix}
}
\description{
Make a new projector instance restricted to a subset of input columns
}
