% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pass}
\alias{pass}
\title{a no-op pre-processing step}
\usage{
pass(preproc = prepper())
}
\arguments{
\item{preproc}{the pre-processing pipeline}
}
\value{
a \code{prepper} list
}
\description{
\code{pass} simply passes its data through the chain
}
