% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{principal components analysis}
\usage{
pca(
  X,
  ncomp = min(dim(X)),
  preproc = center(),
  method = c("base", "fast", "irlba", "propack", "rsvd", "svds"),
  ...
)
}
\arguments{
\item{X}{the data matrix}

\item{ncomp}{the number of requested components to estimate}

\item{preproc}{the pre_processor}

\item{method}{th svd method (passed to \code{svd_wrapper})}

\item{extra}{arguments}
}
\description{
Compute the directions of maximal variance in a matrix via the singula value decomposition
}
\examples{

X <- matrix(rnorm(10*20), 10, 20)
res <- pca(X, ncomp=10, preproc=center())
}
