% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{perm_ci}
\alias{perm_ci}
\title{permutation confidence intervals}
\usage{
perm_ci(x, X, nperm, ...)
}
\arguments{
\item{x}{the model fit}

\item{X}{the original data matrix}

\item{nperm}{the number of permutations}

\item{...}{extra args}
}
\description{
permutation confidence intervals
}
