% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{predict.classifier}
\alias{predict.classifier}
\title{predict with a classifier object}
\usage{
\method{predict}{classifier}(
  object,
  new_data,
  ncomp = NULL,
  colind = NULL,
  metric = c("euclidean", "cosine", "ejaccard"),
  normalize_probs = FALSE,
  ...
)
}
\arguments{
\item{object}{classifier}

\item{new_data}{new data}

\item{ncomp}{number of components}

\item{colind}{column indices}

\item{metric}{similarity metric}

\item{normalize_probs}{logical}

\item{...}{extra args}
}
\value{
list with class and prob
}
\description{
predict with a classifier object
}
