% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{predict.rf_classifier}
\alias{predict.rf_classifier}
\title{predict with a random forest classifier}
\usage{
\method{predict}{rf_classifier}(object, new_data, ncomp = NULL, colind = NULL, ...)
}
\description{
predict with a random forest classifier
}
