% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{prep.prepper}
\alias{prep.prepper}
\title{finalize a prepper pipeline}
\usage{
\method{prep}{prepper}(x, ...)
}
\description{
Prepares a pre-processing pipeline for application by creating \code{init}, \code{transform}, and \code{reverse_transform} functions.
}
