% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{prep}
\alias{prep}
\title{prepare a dataset by applying a pre-processing pipeline}
\usage{
prep(x, ...)
}
\arguments{
\item{x}{the pipeline}

\item{...}{extra args}
}
\value{
the pre-processed data
}
\description{
prepare a dataset by applying a pre-processing pipeline
}
