% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{prep_node}
\alias{prep_node}
\title{prepare a new node and add to pipeline}
\usage{
prep_node(pipeline, name, create, ...)
}
\arguments{
\item{pipeline}{the pre-processing pipeline}

\item{name}{the name of the step to add}

\item{create}{the creation function}
}
\description{
prepare a new node and add to pipeline
}
\keyword{internal}
