% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prinang}
\alias{prinang}
\title{principal angles}
\usage{
prinang(fits)
}
\arguments{
\item{fits}{a list of \code{bi_projector} instances}
}
\value{
a numeric vector of principle angles with length equal to minimum dimension of input subspaces
}
\description{
compute principal angles for a set of subspaces
}
