% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_projector.R
\name{print.bi_projector}
\alias{print.bi_projector}
\title{Pretty Print S3 Method for bi_projector Class}
\usage{
\method{print}{bi_projector}(x, ...)
}
\arguments{
\item{x}{A \code{bi_projector} object}

\item{...}{Additional arguments passed to the print function}
}
\value{
Invisible \code{bi_projector} object
}
\description{
Pretty Print S3 Method for bi_projector Class
}
