% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_projector_union.R
\name{print.bi_projector_union}
\alias{print.bi_projector_union}
\title{Pretty Print S3 Method for bi_projector_union Class}
\usage{
\method{print}{bi_projector_union}(x, ...)
}
\arguments{
\item{x}{A \code{bi_projector_union} object}

\item{...}{Additional arguments passed to the print function}
}
\value{
Invisible \code{bi_projector_union} object
}
\description{
Pretty Print S3 Method for bi_projector_union Class
}
