% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.bootstrap_pca_result}
\alias{print.bootstrap_pca_result}
\title{Print method for bootstrap_pca_result}
\usage{
\method{print}{bootstrap_pca_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bootstrap_pca_result}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print method for bootstrap_pca_result
}
\keyword{internal}
