% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{print.concat_pre_processor}
\alias{print.concat_pre_processor}
\title{Print a concat_pre_processor object}
\usage{
\method{print}{concat_pre_processor}(x, ...)
}
\arguments{
\item{x}{A \code{concat_pre_processor} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a concat_pre_processor object
}
