% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_projector.R
\name{print.perm_test}
\alias{print.perm_test}
\title{Print Method for perm_test Objects}
\usage{
\method{print}{perm_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{perm_test}.}

\item{...}{Additional arguments passed to printing methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a concise summary of the permutation test results.
}
