% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{print.perm_test_pca}
\alias{print.perm_test_pca}
\title{Print Method for perm_test_pca Objects}
\usage{
\method{print}{perm_test_pca}(x, ...)
}
\arguments{
\item{x}{An object of class \code{perm_test_pca}.}

\item{...}{Additional arguments passed to printing methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a concise summary of the PCA permutation test results.
}
