% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{print.pre_processor}
\alias{print.pre_processor}
\title{Print a pre_processor object}
\usage{
\method{print}{pre_processor}(x, ...)
}
\arguments{
\item{x}{A \code{pre_processor} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Display information about a \code{pre_processor} using crayon-based formatting.
}
