% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{print.prepper}
\alias{print.prepper}
\title{Print a prepper pipeline}
\usage{
\method{print}{prepper}(x, ...)
}
\arguments{
\item{x}{A \code{prepper} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Uses \code{crayon} to produce a colorful and readable representation of the pipeline steps.
}
