% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{print.projector}
\alias{print.projector}
\title{Pretty Print Method for \code{projector} Objects}
\usage{
\method{print}{projector}(x, ...)

\method{print}{projector}(x, ...)
}
\arguments{
\item{x}{A \code{projector} object.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
the \code{projector} object
}
\description{
Display a human-readable summary of a \code{projector} object, including information about the dimensions of the projection matrix and the pre-processing pipeline.
}
\examples{
X <- matrix(rnorm(10*10), 10, 10)
svdfit <- svd(X)
p <- projector(svdfit$v)
print(p)
}
