% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{print.rf_classifier}
\alias{print.rf_classifier}
\title{Pretty Print Method for \code{rf_classifier} Objects}
\usage{
\method{print}{rf_classifier}(x, ...)
}
\arguments{
\item{x}{An \code{rf_classifier} object.}

\item{...}{Additional arguments passed to \code{print.randomForest}.}
}
\value{
\code{rf_classifier} object.
}
\description{
Display a human-readable summary of an \code{rf_classifier} object.
}
\examples{
# Assume rf_clf is a fitted rf_classifier object
# print(rf_clf)
}
