% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{project.cross_projector}
\alias{project.cross_projector}
\title{project a cross_projector instance}
\usage{
\method{project}{cross_projector}(x, new_data, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{The model fit, typically an object of class bi_projector or any other class that implements a project method}

\item{new_data}{A matrix or vector of new observations with the same number of columns as the original data. Rows represent observations and columns represent variables}

\item{source}{the source of the data (X or Y block)}

\item{...}{Extra arguments to be passed to the specific project method for the object's class}
}
\value{
the projected data
}
\description{
project a cross_projector instance
}
\seealso{
Other project: 
\code{\link{project}()},
\code{\link{project_block}()},
\code{\link{project_vars}()}
}
\concept{project}
