% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nystrom_embedding.R
\name{project.nystrom_approx}
\alias{project.nystrom_approx}
\title{Project new data using a Nyström approximation model}
\usage{
\method{project}{nystrom_approx}(x, new_data, ...)
}
\arguments{
\item{x}{A \code{nystrom_approx} object (inheriting from \code{bi_projector}).}

\item{new_data}{New data matrix to project.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A matrix of projected scores.
}
\description{
Project new data using a Nyström approximation model
}
