% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project_block}
\alias{project_block}
\title{project a single "block" of data onto the subspace}
\usage{
project_block(x, new_data, block, ...)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of columns as the original data.}

\item{block}{the block id to select in the block projection matrix}

\item{...}{extra args}
}
\description{
When observations are concatenated into "blocks", it may be useful to project one block from the set.
This is equivalent to a "partial projection" where the column indices are associated with a given block.
This is therefore a convenience method of multiblock fits.
}
