% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project_vars}
\alias{project_vars}
\title{new variable projection}
\usage{
project_vars(x, new_data, ...)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of rows as the original data.}

\item{...}{extra args}
}
\description{
project one or more \emph{variables} onto a subspace. This is sometimes called supplementary variable projection and can be computed
for a biorthogonal decomposition such as SVD.
}
