% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{rank_score}
\alias{rank_score}
\title{Calculate Rank Score for Predictions}
\usage{
rank_score(prob, observed)
}
\arguments{
\item{prob}{matrix of predicted probabilities (observations x classes)}

\item{observed}{vector of observed class labels}
}
\value{
data.frame with prank and observed
}
\description{
Calculate Rank Score for Predictions
}
