% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstruct.pca}
\alias{reconstruct.pca}
\title{Reconstruct Data from PCA Results}
\usage{
\method{reconstruct}{pca}(x, comp = 1:ncomp(x), ...)
}
\arguments{
\item{x}{A \code{pca} object.}

\item{comp}{Integer vector specifying which components to use for reconstruction (default: all components in \code{x}).}

\item{...}{Extra arguments (ignored).}
}
\value{
A matrix representing the reconstructed data in the \emph{original} scale (preprocessing reversed).
}
\description{
Reconstructs the original (centered) data matrix from the PCA scores and loadings.
}
