% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reconstruct}
\alias{reconstruct}
\title{reconstruct the data}
\usage{
reconstruct(x, comp, rowind, colind, ...)
}
\arguments{
\item{x}{the model fit}

\item{comp}{a vector of component indices to use in the reconstruction.}

\item{rowind}{the row indices to reconstruct}

\item{colind}{the column indices to reconstruct}

\item{...}{extra args}
}
\description{
Reconstruct a data set from its (possibly) low-rank representation.
}
