% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{refit}
\alias{refit}
\title{refit a model}
\usage{
refit(x, new_data, ...)
}
\arguments{
\item{x}{the original model fit object}

\item{new_data}{the new data to process}

\item{...}{extra args}
}
\value{
a refit model object
}
\description{
refit a model given new data or new parameter(s)
}
