% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{reprocess.cross_projector}
\alias{reprocess.cross_projector}
\title{reprocess a cross_projector instance}
\usage{
\method{reprocess}{cross_projector}(x, new_data, colind = NULL, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{the model fit object}

\item{new_data}{the new data to process}

\item{colind}{the column indices of the new data}

\item{source}{the source of the data (X or Y block)}

\item{...}{extra args}
}
\value{
the re(pre-)processed data
}
\description{
reprocess a cross_projector instance
}
\concept{reprocess}
