% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{residuals}
\alias{residuals}
\title{get residuals of component model fit}
\usage{
residuals(x, ncomp, xorig, ...)
}
\arguments{
\item{x}{the model fit}

\item{ncomp}{the number of components to factor out}

\item{xorig}{the original X data matrix}

\item{...}{extra arguments}
}
\description{
Extract the residuals of a model, after removing the first \code{ncomp} components
}
