% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reverse_transform}
\alias{reverse_transform}
\title{reverse a pre-processing transform}
\usage{
reverse_transform(x, X, colind, ...)
}
\arguments{
\item{x}{the pre_processor}

\item{X}{the data matrix}

\item{colind}{column indices}

\item{...}{extra args}
}
\value{
the reverse-transformed data
}
\description{
reverse a pre-processing transform
}
