% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{rf_classifier.projector}
\alias{rf_classifier.projector}
\title{Create a random forest classifier}
\usage{
\method{rf_classifier}{projector}(x, colind = NULL, labels, scores, ...)
}
\arguments{
\item{x}{a projector object}

\item{colind}{optional col indices}

\item{labels}{class labels}

\item{scores}{reference scores}

\item{...}{passed to \code{randomForest}}
}
\value{
a \code{rf_classifier} object with rfres (rf model), labels, scores
}
\description{
Uses \code{randomForest} to train a random forest on the provided scores and labels.
}
