% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{robust_inv_vTv}
\alias{robust_inv_vTv}
\title{Possibly use ridge-regularized inversion of crossprod(v)}
\usage{
robust_inv_vTv(v, lambda = 1e-06)
}
\description{
Possibly use ridge-regularized inversion of crossprod(v)
}
\keyword{internal}
