% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{rotate}
\alias{rotate}
\title{Rotate a Component Solution}
\usage{
rotate(x, ncomp, type, ...)
}
\arguments{
\item{x}{The model fit, typically a result from a dimensionality reduction method like PCA.}

\item{ncomp}{The number of components to rotate.}

\item{type}{The type of rotation to apply (e.g., "varimax", "quartimax", "promax").}

\item{...}{extra args}
}
\value{
A modified model fit with the rotated components.
}
\description{
Perform a rotation of the component loadings to improve interpretability.
}
