% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{scores.cca}
\alias{scores.cca}
\title{Extract scores from a CCA fit}
\usage{
\method{scores}{cca}(x, block = c("X", "Y"), ...)
}
\arguments{
\item{x}{A \code{cca} object.}

\item{block}{Which block to return scores for: "X" (default) or "Y".}

\item{...}{Ignored.}
}
\value{
Numeric matrix of canonical scores for the chosen block.
}
\description{
Extract scores from a CCA fit
}
