% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsc.R
\name{scores.plsc}
\alias{scores.plsc}
\title{Extract scores from a PLSC fit}
\usage{
\method{scores}{plsc}(x, block = c("X", "Y"), ...)
}
\arguments{
\item{x}{A \code{plsc} object.}

\item{block}{Which block to return scores for: "X" (default) or "Y".}

\item{...}{Ignored.}
}
\value{
Numeric matrix of scores for the chosen block.
}
\description{
Extract scores from a PLSC fit
}
