% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{scores}
\alias{scores}
\title{get the component scores}
\usage{
scores(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
Extract the factor score matrix from a fit. These are the projections of the data onto the components.
}
