% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{screeplot.pca}
\alias{screeplot.pca}
\title{Screeplot for PCA}
\usage{
\method{screeplot}{pca}(x, type = "barplot", main = "Screeplot", ...)
}
\arguments{
\item{x}{A \code{pca} object.}

\item{type}{"barplot" or "lines".}

\item{main}{Plot title.}

\item{...}{Additional args to pass to base R plotting.}
}
\description{
Displays the variance explained by each principal component as a bar or line plot.
}
