% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{screeplot}
\alias{screeplot}
\title{Screeplot for PCA}
\usage{
screeplot(x, ...)
}
\arguments{
\item{x}{A \code{pca} object.}

\item{...}{extra args}
}
\description{
Displays the variance explained by each principal component as a bar or line plot.
}
