% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{sdev}
\alias{sdev}
\title{standard deviations}
\usage{
sdev(x)
}
\arguments{
\item{x}{the model fit}
}
\value{
the standard deviations
}
\description{
The standard deviations of the projected data matrix
}
