% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{shape.cross_projector}
\alias{shape.cross_projector}
\title{shape of a cross_projector instance}
\usage{
\method{shape}{cross_projector}(x, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{The model fit.}

\item{source}{the source of the data (X or Y block)}

\item{...}{Extra arguments.}
}
\value{
the shape of the data
}
\description{
shape of a cross_projector instance
}
\concept{shape}
