% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/twoway_projector.R
\name{shape}
\alias{shape}
\alias{shape.cross_projector}
\title{Shape of the Projector}
\usage{
shape(x, ...)

\method{shape}{cross_projector}(x, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{The model fit.}

\item{...}{Extra arguments.}

\item{source}{the source of the data (X or Y block)}
}
\value{
A vector containing the dimensions of the sample loadings matrix \code{v} (number of rows and columns).
}
\description{
Get the input/output shape of the projector.
}
\details{
This function retrieves the dimensions of the sample loadings matrix \code{v} in the form of a vector with two elements.
The first element is the number of rows in the \code{v} matrix, and the second element is the number of columns.
}
