% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_utils.R
\name{shared_pca}
\alias{shared_pca}
\title{Shared PCA basis specification}
\usage{
shared_pca(ncomp, fit_on = c("nuisance_residual", "full", "whitened_residual"))
}
\arguments{
\item{ncomp}{Number of shared basis components.}

\item{fit_on}{Basis fitting target.}
}
\value{
A basis specification for a shared PCA basis.
}
\description{
Shared PCA basis specification
}
