% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{standardize}
\alias{standardize}
\title{center and scale each vector of a matrix}
\usage{
standardize(preproc = prepper(), cmeans = NULL, sds = NULL)
}
\arguments{
\item{preproc}{the pre-processing pipeline}

\item{cmeans}{an optional vector of column means}

\item{sds}{an optional vector of sds}
}
\value{
a \code{prepper} list
}
\description{
center and scale each vector of a matrix
}
