% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{std_scores}
\alias{std_scores}
\title{compute standardized scores}
\usage{
std_scores(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
compute standardized scores
}
