% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subspace_similarity}
\alias{subspace_similarity}
\title{Compute subspace similarity}
\usage{
subspace_similarity(
  fits,
  method = c("avg_pair", "grassmann", "worst_case"),
  ...
)
}
\arguments{
\item{fits}{a list of bi_projector objects}

\item{method}{the method to use for computing subspace similarity}

\item{...}{additional arguments to pass to the method}
}
\value{
a numeric value representing the subspace similarity
}
\description{
Compute subspace similarity
}
