% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{summarize_by}
\alias{summarize_by}
\title{summarize data over grouping variable(s)}
\usage{
summarize_by(x, ..., sfun)
}
\arguments{
\item{x}{the multivariate data (a matrix, a list, or other data container)}

\item{...}{the grouping variables}

\item{sfun}{the columnwise summary function (e.g. \code{colMeans})}
}
\description{
apply a mutlivariate columnwise summary function (e.g. \code{colMeans}) to subsets
of observations formed by one or more grouping variables.
}
