% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{svd_wrapper}
\alias{svd_wrapper}
\title{svd_wrapper}
\usage{
svd_wrapper(
  X,
  ncomp = min(dim(X)),
  preproc = pass(),
  method = c("base", "fast", "irlba", "propack", "rsvd", "svds"),
  q = 2,
  p = 10,
  tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{X}{the \code{matrix}}

\item{ncomp}{number of components to estimate}

\item{method}{the svd method to use. One of: 'base', 'fast', 'irlba', 'propack', 'rsvd', 'svds'}

\item{q}{parameter passed to method \code{rsvd}}

\item{p}{parameter passed to method \code{rsvd}}

\item{tol}{minimum eigenvalue magnitude, otherwise component is dropped}
}
\value{

}
\description{
svd_wrapper
}
\examples{

data(iris)
X <- iris[,1:4]

fit <- svd_wrapper(X, ncomp=3, preproc=center(), method="base")


an `svd` object that extends `projector`
}
