% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{transpose}
\alias{transpose}
\title{transpose a model}
\usage{
transpose(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
transpose a model by switching coefficients and scores
}
