% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed_projector.R
\name{truncate.composed_projector}
\alias{truncate.composed_projector}
\title{Truncate a Composed Projector}
\usage{
\method{truncate}{composed_projector}(x, ncomp, ...)
}
\arguments{
\item{x}{A \code{composed_projector} object.}

\item{ncomp}{The desired number of final output components.}

\item{...}{Currently unused.}
}
\value{
A new \code{composed_projector} object with the last stage truncated.
}
\description{
Reduces the number of output components of the composed projector by
truncating the \emph{last} stage in the sequence.
}
\details{
Note: This implementation currently only supports truncating the final stage.
Truncating intermediate stages would require re-computing subsequent stages
or combined attributes and is not yet implemented.
}
