% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{truncate}
\alias{truncate}
\title{truncate a component fit}
\usage{
truncate(x, ncomp)
}
\arguments{
\item{x}{the object to truncate}

\item{ncomp}{number of components to retain}
}
\description{
take the first n components of a decomposition
}
