% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{twoway_projector}
\alias{twoway_projector}
\title{Two-way projection to latent components}
\usage{
twoway_projector(
  vx,
  vy,
  preproc_x = prep(pass()),
  preproc_y = prep(pass()),
  ...,
  classes = NULL
)
}
\arguments{
\item{vx}{the X coefficients}

\item{vy}{the Y coefficients}

\item{preproc_x}{the X pre-processor}

\item{preproc_y}{the Y pre-processor}

\item{...}{extra parameters or results to store}

\item{classes}{additional class names}
}
\description{
A projector that reduces two blocks of data, X and Y, yielding a pairs of weights for each component.
This structure can be used for example to store weights derived from canonical correlation analysis.
}
\details{
This class extends \code{projector} and therefore basic operations such as \code{project}, \code{shape}, \code{reprocess},
and \code{coef} work, but by default it is assumed that the \code{X} block is primary. To access \code{Y} block operations, an
additional argument \code{source} must be supplied to the relevant functions, e.g. \code{coef(fit, source="Y")}
}
