% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{apply_rotation}
\alias{apply_rotation}
\title{Apply rotation}
\usage{
apply_rotation(x, rotation_matrix, ...)
}
\arguments{
\item{x}{A model object, possibly created using the \code{pca()} function.}

\item{rotation_matrix}{\code{matrix} reprsenting the rotation.}

\item{...}{extra args}
}
\value{
A modified object with updated components and scores after applying the specified rotation.
}
\description{
Apply a specified rotation to the fitted model
}
