% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_projector.R
\name{bi_projector}
\alias{bi_projector}
\title{construct a \code{bi_projector} instance}
\usage{
bi_projector(v, s, sdev, preproc = prep(pass()), classes = NULL, ...)
}
\arguments{
\item{v}{a matrix of coefficients with dimension \code{nrow(v)} by \code{ncol(v)} (number of columns = number of components)}

\item{s}{the score matrix}

\item{sdev}{the standard deviations of th score matrix}

\item{preproc}{a prepped pre-processing object (default is the no-processing \code{pass()} pre_processor)}

\item{classes}{additional class information used for creating subtypes of \code{projector}}

\item{...}{extra args}
}
\description{
A \code{bi_projector} offers a two-way mapping from from samples (rows) to scores and from variables (columns) to components.
Thus, one can project from D-dimensional input space to d-dimensional subspace. And one can project (\code{project_vars}) from n-dimensional
variable space to the d-dimensional component space.
}
