% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bootstrap_pca.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.pca}
\title{bootstrap resampling}
\usage{
bootstrap(x, nboot, ...)

\method{bootstrap}{pca}(x, nboot = 100, k = ncomp(x))
}
\arguments{
\item{x}{the model fit}

\item{nboot}{the number of bootstrap resamples}

\item{...}{extra args}
}
\value{
A list with the bootstrap z-scores for the loadings (\code{zboot_loadings}) and scores (\code{zboot_scores})
}
\description{
bootstrap a multivariate model to estimate component and score variability
}
\examples{

X <- matrix(rnorm(10*100), 10, 100)
x = pca(X, ncomp=9)


}
\references{
Fisher, Aaron, Brian Caffo, Brian Schwartz, and Vadim Zipunnikov. 2016.
“Fast, Exact Bootstrap Principal Component Analysis for P > 1 Million.” \emph{Journal of the American Statistical Association} 111 (514): 846–60.
}
