% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/plsc_inference.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.plsc}
\title{Bootstrap Resampling for Multivariate Models}
\usage{
bootstrap(x, nboot, ...)

\method{bootstrap}{plsc}(x, nboot = 500, ...)
}
\arguments{
\item{x}{A fitted model object, such as a \code{projector}, that has been fit to a training dataset.}

\item{nboot}{An integer specifying the number of bootstrap resamples to perform.}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{bootstrap}.}
}
\value{
A list containing the bootstrap resampled components and scores for the model.
}
\description{
Perform bootstrap resampling on a multivariate model to estimate the variability of components and scores.
}
