% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.discriminant_projector}
\alias{classifier.discriminant_projector}
\title{Create a k-NN classifier for a discriminant projector}
\usage{
\method{classifier}{discriminant_projector}(x, colind = NULL, knn = 1, ...)
}
\arguments{
\item{x}{the discriminant projector object}

\item{colind}{an optional vector specifying the column indices of the components to use for prediction (NULL by default)}

\item{knn}{the number of nearest neighbors to consider in the k-NN classifier (default is 1)}

\item{...}{extra arguments}
}
\value{
a classifier object
}
\description{
Constructs a k-NN classifier for a discriminant projector, with an option to use a subset of the components.
}
