% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.multiblock_biprojector}
\alias{classifier.multiblock_biprojector}
\title{Multiblock Bi-Projector Classifier}
\usage{
\method{classifier}{multiblock_biprojector}(
  x,
  colind = NULL,
  labels,
  new_data = NULL,
  block = NULL,
  knn = 1,
  ...
)
}
\arguments{
\item{x}{A fitted multiblock bi-projector model object.}

\item{colind}{An optional vector of column indices used for prediction (default: NULL).}

\item{labels}{A factor or vector of class labels for the training data.}

\item{new_data}{An optional data matrix for which to generate predictions (default: NULL).}

\item{block}{An optional block index for prediction (default: NULL).}

\item{knn}{The number of nearest neighbors to consider in the classifier (default: 1).}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{classifier}.}
}
\value{
A multiblock classifier object.
}
\description{
Constructs a classifier for a multiblock bi-projector model that can generate predictions for new data points.
}
\seealso{
Other classifier: 
\code{\link{classifier.projector}()}
}
\concept{classifier}
