% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.projector}
\alias{classifier.projector}
\title{create classifier from a projector}
\usage{
\method{classifier}{projector}(
  x,
  colind = NULL,
  labels,
  new_data = NULL,
  knn = 1,
  global_scores = TRUE,
  ...
)
}
\arguments{
\item{x}{projector}

\item{colind}{...}

\item{labels}{...}

\item{new_data}{...}

\item{knn}{...}

\item{global_scores}{...}

\item{...}{extra args}
}
\description{
create classifier from a projector
}
\examples{
# Assume proj is a fitted projector object
# Assume lbls are labels and dat is new data
# classifier(proj, labels = lbls, new_data = dat, knn = 3)
}
\seealso{
Other classifier: 
\code{\link{classifier.multiblock_biprojector}()},
\code{\link{rf_classifier.projector}()}
}
\concept{classifier}
