% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier.projector}
\alias{classifier.projector}
\title{create a classifier}
\usage{
\method{classifier}{projector}(x, labels, new_data, colind = NULL, knn = 1)
}
\arguments{
\item{x}{the \code{projector} instance}

\item{labels}{the labels associated with the rows of the projected data (see \code{new_data})}

\item{new_data}{reference data associated with \code{labels} and to be projected into subspace (required).}

\item{colind}{the subset of column indices in the fitted model to use.}

\item{knn}{the number of nearest neighbors to use when classifying a new point.}
}
\description{
construct a classifier from a \code{projector} instance
}
\examples{
data(iris)
X <- iris[,1:4]
pcres <- pca(as.matrix(X),2)
cfier <- classifier(pcres, iris[,5])
}
