% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{classifier}
\alias{classifier}
\title{construct a classifier}
\usage{
classifier(x, colind, ...)
}
\arguments{
\item{x}{the model object}

\item{colind}{the (optional) column indices used for prediction}
}
\description{
Given a model object (e.g. \code{projector} construct a classifier that can generate predictions for new data points.
}
