% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{classifier}
\alias{classifier}
\title{Construct a Classifier}
\usage{
classifier(x, colind, ...)
}
\arguments{
\item{x}{A model object, such as a \code{projector}, that has been fit to a training dataset.}

\item{colind}{Optional vector of column indices used for prediction. If not provided, all columns will be used.}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{classifier}.}
}
\value{
A classifier function that can be used to make predictions on new data points.
}
\description{
Create a classifier from a given model object (e.g., \code{projector}). This classifier can generate predictions for new data points.
}
