% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/classifier.R
\name{classifier}
\alias{classifier}
\alias{classifier.projector}
\title{Construct a Classifier}
\usage{
classifier(x, colind, ...)

\method{classifier}{projector}(
  x,
  colind = NULL,
  labels,
  new_data = NULL,
  knn = 1,
  global_scores = TRUE,
  ...
)
}
\arguments{
\item{x}{projector}

\item{colind}{...}

\item{...}{extra args}

\item{labels}{...}

\item{new_data}{...}

\item{knn}{...}

\item{global_scores}{...}
}
\value{
A classifier function that can be used to make predictions on new data points.
}
\description{
Create a classifier from a given model object (e.g., \code{projector}). This classifier can generate predictions for new data points.
}
\examples{
# Assume proj is a fitted projector object
# Assume lbls are labels and dat is new data
# classifier(proj, labels = lbls, new_data = dat, knn = 3)
}
\seealso{
Other classifier: 
\code{\link{classifier.multiblock_biprojector}()},
\code{\link{rf_classifier.projector}()}
}
\concept{classifier}
