% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway_projector.R
\name{coef.cross_projector}
\alias{coef.cross_projector}
\title{Extract coefficients from a cross_projector object}
\usage{
\method{coef}{cross_projector}(object, source = c("X", "Y"), ...)
}
\arguments{
\item{object}{the model fit}

\item{source}{the source of the data (X or Y block), either "X" or "Y"}

\item{...}{extra args}
}
\value{
the coefficients
}
\description{
Extract coefficients from a cross_projector object
}
