% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{colscale}
\alias{colscale}
\title{scale a data matrix}
\usage{
colscale(preproc = prepper(), type = c("unit", "z", "weights"), weights = NULL)
}
\arguments{
\item{preproc}{the pre-processing pipeline}

\item{type}{the kind of scaling, \code{unit} norm, \code{z}-scoring, or precomputed \code{weights}}

\item{weights}{optional precomputed weights}
}
\value{
a \code{prepper} list
}
\description{
normalize each column by a scale factor.
}
