% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{components}
\alias{components}
\title{get the components}
\usage{
components(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\value{
the component matrix
}
\description{
Extract the component matrix of a fit.
}
