% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{compose_projector}
\alias{compose_projector}
\title{compose two projectors}
\usage{
compose_projector(x, y, ...)
}
\arguments{
\item{x}{the first projector}

\item{y}{the second projector}

\item{...}{extra args}
}
\description{
compose two projectors
}
