% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_projector.R
\name{discriminant_projector}
\alias{discriminant_projector}
\title{construct a discriminant projector}
\usage{
discriminant_projector(
  v,
  s,
  sdev,
  preproc = prep(pass()),
  labels,
  classes = NULL,
  ...
)
}
\arguments{
\item{v}{a matrix of coefficients with dimension \code{nrow(v)} by \code{ncol(v)} (number of columns = number of components)}

\item{s}{the score matrix}

\item{sdev}{the standard deviations of th score matrix}

\item{preproc}{a prepped pre-processing object (default is the no-processing \code{pass()} pre_processor)}

\item{labels}{the training labels}

\item{classes}{additional class information used for creating subtypes of \code{projector}}

\item{...}{extra args}
}
\description{
a \code{discriminant_projector} instance extending \code{bi_projeector} whose projection maximizes class separation.
}
