% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{inverse_projection}
\alias{inverse_projection}
\title{Inverse of the Component Matrix}
\usage{
inverse_projection(x, ...)
}
\arguments{
\item{x}{The model fit.}

\item{...}{Extra arguments.}
}
\value{
The inverse projection matrix.
}
\description{
Return the inverse projection matrix, which can be used to map back to data space.
If the component matrix is orthogonal, then the inverse projection is the transpose of the component matrix.
}
\seealso{
\code{\link{project}} for projecting data onto the subspace.
}
