% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{multiblock_biprojector}
\alias{multiblock_biprojector}
\title{Create a multiblock bi-projector}
\usage{
multiblock_biprojector(
  v,
  s,
  sdev,
  preproc = prep(pass()),
  ...,
  block_indices,
  classes = NULL
)
}
\arguments{
\item{v}{the component matrix}

\item{s}{the score matrix}

\item{sdev}{the singular values}

\item{preproc}{a function for preprocessing the data (default is a pass-through)}

\item{...}{extra arguments}

\item{block_indices}{a list of numeric vectors specifying the indices of each data block}

\item{classes}{additional S3 classes to assign to the bi-projector object}
}
\value{
a multiblock bi-projector object
}
\description{
Constructs a multiblock bi-projector using the given component matrix (v), score matrix (s), singular values (sdev), a preprocessing function, and a list of block indices.
}
