% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{multiblock_projector}
\alias{multiblock_projector}
\title{construct a multiblock project}
\usage{
multiblock_projector(
  v,
  preproc = prep(pass()),
  ...,
  block_indices,
  classes = NULL
)
}
\arguments{
\item{v}{a matrix of coefficients with dimension \code{nrow(v)} by \code{ncol(v)} (number of columns = number of components)}

\item{preproc}{a prepped pre-processing object (default is the no-processing \code{pass()} pre_processor)}

\item{...}{extra args}

\item{block_indices}{the list of indices indices of the data blocks}

\item{classes}{additional class information used for creating subtypes of \code{projector}}
}
\description{
construct a multiblock project
}
