% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{new_classifier}
\alias{new_classifier}
\title{Create a new k-NN classifier for a model fit}
\usage{
new_classifier(x, labels, scores, colind = NULL, knn = 1, classes = NULL, ...)
}
\arguments{
\item{x}{the model fit}

\item{labels}{a vector of class labels corresponding to the data points}

\item{scores}{a matrix of scores used for classification}

\item{colind}{an optional vector specifying the column indices of the components to use for prediction (NULL by default)}

\item{knn}{the number of nearest neighbors to consider in the k-NN classifier (default is 1)}

\item{classes}{additional S3 classes to assign to the classifier object}

\item{...}{extra arguments}
}
\value{
a classifier object
}
\description{
Constructs a new k-NN classifier based on a model fit and the corresponding scores matrix, with an option to use a subset of the components.
}
\keyword{internal}
