% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{new_classifier}
\alias{new_classifier}
\title{construct a new classifier}
\usage{
new_classifier(x, labels, scores, colind = NULL, knn = 1, classes = NULL, ...)
}
\arguments{
\item{x}{the model fit}

\item{labels}{the class labels}

\item{colind}{the column indices}

\item{knn}{the number of nearest neighbors for prediction}

\item{classes}{addiitonal s3 classes}

\item{...}{extra args}
}
\description{
construct a new classifier
}
\examples{

X <- matrix(rnorm(10*10), 10,10)
L <- letters[1:10]
pc <- pca(X, ncomp=8)

cf <- new_classifier(pc, L, scores(pc), colind=2:7)
predict(cf, X[,2:7], ncomp=3, metric="cosine")
}
\keyword{internal}
