% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_inverse_projection}
\alias{partial_inverse_projection}
\title{Partial Inverse Projection of a Columnwise Subset of Component Matrix}
\usage{
partial_inverse_projection(x, colind, ...)
}
\arguments{
\item{x}{A fitted model object, such as a \code{projector}, that has been fit to a dataset.}

\item{colind}{A numeric vector specifying the column indices of the component matrix to consider for the partial inverse projection.}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{partial_inverse_projection}.}
}
\value{
A matrix representing the partial inverse projection.
}
\description{
Compute the inverse projection of a columnwise subset of the component matrix (e.g., a sub-block).
Even when the full component matrix is orthogonal, there is no guarantee that the partial component matrix is orthogonal.
}
