% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{principal components analysis}
\usage{
pca(
  X,
  ncomp = min(dim(X)),
  preproc = center(),
  method = c("fast", "base", "irlba", "propack", "rsvd", "svds"),
  ...
)
}
\arguments{
\item{X}{the data matrix}

\item{ncomp}{the number of requested components to estimate}

\item{preproc}{the pre_processor}

\item{method}{th svd method (passed to \code{svd_wrapper})}

\item{extra}{arguments}
}
\description{
Compute the directions of maximal variance in a matrix via the singula value decomposition
}
\examples{

data(iris)
X <- as.matrix(iris[,1:4])
res <- pca(X, ncomp=4)
tres <- truncate(res, 3)
}
