% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{perm_ci}
\alias{perm_ci}
\title{Permutation Confidence Intervals}
\usage{
perm_ci(x, X, nperm, ...)
}
\arguments{
\item{x}{A model fit object.}

\item{X}{The original data matrix used to fit the model.}

\item{nperm}{The number of permutations to perform for the confidence interval estimation.}

\item{...}{Additional arguments to be passed to the specific model implementation of \code{perm_ci}.}
}
\value{
A list containing the estimated lower and upper bounds of the confidence intervals for model parameters.
}
\description{
Estimate confidence intervals for model parameters using permutation testing.
}
