% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{predict.classifier}
\alias{predict.classifier}
\title{predict with a classifier object}
\usage{
\method{predict}{classifier}(
  object,
  new_data,
  ncomp = NULL,
  colind = NULL,
  metric = c("cosine", "euclidean"),
  ...
)
}
\arguments{
\item{object}{the model fit}

\item{new_data}{new data to predict on}

\item{ncomp}{the number of components to use}

\item{colind}{the column indices to select in the projection matrix}

\item{metric}{the similarity metric ("euclidean" or "cosine")}

\item{...}{additional arguments to projection function}
}
\value{
a list with the predicted class and probabilities
}
\description{
predict with a classifier object
}
