% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{print.classifier}
\alias{print.classifier}
\title{Pretty Print Method for \code{classifier} Objects}
\usage{
\method{print}{classifier}(x, ...)
}
\arguments{
\item{x}{A \code{classifier} object.}

\item{...}{Additional arguments.}
}
\value{
\code{classifier} object.
}
\description{
Display a human-readable summary of a \code{classifier} object.
}
\examples{
# Assume clf is a fitted classifier object
# print(clf)
}
