% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{print.multiblock_biprojector}
\alias{print.multiblock_biprojector}
\title{Pretty Print Method for \code{multiblock_biprojector} Objects}
\usage{
\method{print}{multiblock_biprojector}(x, ...)
}
\arguments{
\item{x}{A \code{multiblock_biprojector} object.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
Invisible \code{multiblock_biprojector} object.
}
\description{
Display a summary of a \code{multiblock_biprojector} object.
}
