% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project}
\alias{project}
\title{new sample projection}
\usage{
project(x, new_data, ...)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of columns as the original data.}

\item{...}{extra args}
}
\description{
Project one or more \emph{samples} of onto a subspace.
}
