% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/twoway_projector.R
\name{project}
\alias{project}
\alias{project.twoway_projector}
\title{new sample projection}
\usage{
project(x, new_data, ...)

\method{project}{twoway_projector}(x, new_data, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{the model fit}

\item{new_data}{a matrix or vector of new observations(s) with the same number of columns as the original data.}

\item{...}{extra args}

\item{source}{the source of the data (X or Y block)}
}
\description{
Project one or more \emph{samples} of onto a subspace.
}
