% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projector.R
\name{projector}
\alias{projector}
\title{construct a \code{projector} instance}
\usage{
projector(v, preproc = NULL, ..., classes = NULL)
}
\arguments{
\item{v}{a matrix of coefficients with dimension \code{nrow(v)} by \code{ncol(v)} (number of columns = number of components)}

\item{preproc}{a pre-processing object}

\item{classes}{additional class information used for creating subtypes of \code{projector}}
}
\value{
a instance of type \code{projector}
}
\description{
A \code{projector} maps a matrix from D-dimensional space to d-dimensional space, where \code{d} may be less than \code{D}.
The projection matrix, \code{v} is not necessarily orthogonal.
}
