% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Multi-output linear regression}
\usage{
regress(
  X,
  Y,
  preproc = NULL,
  method = c("lm", "ridge", "pls"),
  intercept = FALSE,
  lambda = 0.001,
  ncomp = ceiling(ncol(X)/2)
)
}
\arguments{
\item{X}{the set of independent (basis) variables}

\item{Y}{the response matrix}

\item{preproc}{the pre-processor}

\item{method}{the regression method: \code{linear} or \code{ridge}.}
}
\value{
an \code{bi-projector} of type \code{regress}
}
\description{
Fit a multivariate regression model for a matrix of basis functions, \code{X}, and a response matrix \code{Y}.
The goal is to find a projection matrix that can be used for mapping and reconstruction.
}
\examples{

Y <- matrix(rnorm(100*10), 10, 100)
X <- matrix(rnorm(10*9), 10, 9)
r <- regress(X,Y, intercept=FALSE)
recon <- reconstruct(r)
r <- regress(X,Y, intercept=TRUE)
recon <- reconstruct(r)
r <- regress(X,Y, intercept=TRUE, method="ridge")
recon <- reconstruct(r)
}
