% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{residuals}
\alias{residuals}
\title{Obtain residuals of a component model fit}
\usage{
residuals(x, ncomp, xorig, ...)
}
\arguments{
\item{x}{The model fit object.}

\item{ncomp}{The number of components to factor out before calculating residuals.}

\item{xorig}{The original data matrix (X) used to fit the model.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A matrix of residuals, with the same dimensions as the original data matrix.
}
\description{
Calculate the residuals of a model after removing the effect of the first \code{ncomp} components.
This function is useful to assess the quality of the fit or to identify patterns that are not
captured by the model.
}
\concept{residuals}
