% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{rf_classifier.projector}
\alias{rf_classifier.projector}
\title{create a random forest classifier}
\usage{
\method{rf_classifier}{projector}(x, labels, scores, colind = NULL, ...)
}
\description{
create a random forest classifier
}
\examples{
data(iris)
X <- iris[,1:4]
pcres <- pca(as.matrix(X),2)
cfier <- rf_classifier(pcres, labels=iris[,5], scores=scores(pcres))
p <- predict(cfier, as.matrix(iris[,1:4]))
}
