% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{rf_classifier.projector}
\alias{rf_classifier.projector}
\title{create a random forest classifier}
\usage{
\method{rf_classifier}{projector}(x, colind = NULL, labels, scores, ...)
}
\arguments{
\item{x}{the model object}

\item{colind}{the (optional) column indices used for prediction}

\item{labels}{A factor or vector of class labels for the training data.}

\item{scores}{a matrix of references scores used for classification}

\item{...}{extra arguments to \code{randomForest} function}
}
\value{
a \code{rf_classifier} object
}
\description{
create a random forest classifier
}
\examples{
data(iris)
X <- iris[,1:4]
pcres <- pca(as.matrix(X),2)
cfier <- rf_classifier(pcres, labels=iris[,5], scores=scores(pcres))
p <- predict(cfier, new_data=as.matrix(iris[,1:4]))
}
