% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/twoway_projector.R
\name{shape}
\alias{shape}
\alias{shape.cross_projector}
\title{shape of the projector}
\usage{
shape(x, ...)

\method{shape}{cross_projector}(x, source = c("X", "Y"), ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}

\item{source}{the source of the data (X or Y block)}
}
\description{
Get the input/output shape of the projector
}
